Display map route on map view in walk detail view Walk detail view presented once walk has been saved in track walk view, passing coordinates explicitly If detail walk view is not loaded with explicit walk, it retrieves a walk from the API given a walk ID

+++ Amble/Walk.swift
@@ -13,7 +13,7 @@ struct Walk {
   
   let name: String
   let coordinates: [CLLocationCoordinate2D]
-  let time: Double
+  let time: Int
   let distance: Double
   let calories: Double
